Add the OfxSetBundleDirectory API point
This is an addition to the two symbols that must be defined by plugin
binaries. Since it is not part of Core OpenFX, it is not mandatory,
hence hosts must not reject the binary when it is not implemented.

I found this unavaidable to implement binaries for which plugins depend
on manifest and data located in the bundle resources, so that the
behavior of the plugin can be edited without rebuilding the binaries.
Since the number of plugins in the binary itself may be influenced, the
OfxSetBundleDirectory symbol must be called first, if ever.