Breaking change - Add attribute semantics and attribute request mechanism
Attribute semantics tell what the intended usage of the attribute data
is prior to and beyond the effect.

Attribute request is a mean for the effect to tell the host at describe
time that it will need a particular attribute to be available at cook
time. It can also notify the host from attributes that are not mandatory
but that it would make good use of, to have the host recook the effect
any time this attribute changes.